reduce retries in staking:
Staking code unnecessarily causes optimistic execution to check for exact balance matches on the staking contract account when all it needs is to check a lower bound on a balance.  This PR fixes that and should result in fewer retries/re-execs: below is a minimal concrete scenario showing avoidance of a retry. Although not illustrated by this minimal example, not needing to retry a tx can snowball into more performance improvements as the later txs will become less likely to see a stale state and thus needing retries.

  Pre-state

  - Validators 1 and 2 already exist
  - STAKING_CA holds 100 MON.
  - Delegator of validator 2 has a pending withdrawal request for 10 MON.

  Transactions

  1. external_reward(val_id = 1, msg.value = 20 MON)
  2. withdraw(val_id = 2, withdrawal_id = 0) for 10 MON

Consider the following interleaving:

1. Tx2  reads `STAKING_CA.balance` for the `ASSERT`. before this PR, this records the assumption `STAKING_CA.balance==100`. After this PR, the recorded assumption `STAKING_CA.balance>=20`. (Because insolvency is an invariant, we actually do not need to record even this assumption, but that would require breaking some abstractions desirable to be enforced in `State` )
2. Then Tx1 finishes updating `STAKING.CA.balance` to 120
3. Tx1 commits (`BlockState.merge`)
4. Tx2 tries to commit. before this PR,  `BlockState.can_merge` retrurns false because the `STAKING_CA.balance==100`  assumption is violated. So tx2 will be unnecessarily retried. After the PR, tx1 will succeed.

This assumes that there is no persistent storage slot that Tx1 updates and Tx2 reads (this is why the 2 Txs target different validators), and that the nonce and code hash fields  of the staking contract never change.